+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Changes to the Lean backend:
  - Add generation of specs from requires/ensures-annotations (#1815)
  - Add support for nonliteral array sizes (#1826)
  - Add `hax_lib::lean::proof` attribute (#1831)
+ - Add support for `#[hax_lib::opaque]` (#1846)
 
 Miscellaneous:
 - Reserve extraction folder for auto-generated files in Lean examples (#1754)